Windows port: full test/lint/type-check pass, native path handling, Go/Python test portability
- Path normalization in launcher/utils.h understands drive letters, UNC roots and backslashes
- Native owner-SID check for shared memory (fd_owned_by_current_user), WTS based num_users
- ty type-check runs on Windows: target linux platform in pyproject, pass MSYS2 site-packages
- Skip X11 startup notification on Windows
- Go tests use filepath.FromSlash for platform separators; nested 3rdparty/go-shm module excluded from the main module's test discovery
- Completion, file transmission, DND, graphics fd counting and pty tests ported to Windows
- autoformat applied to the win32 GLFW sources now compiled into kitty

+++ kitty/disk-cache.c
@@ -1100,7 +1100,7 @@ get(PyObject *self, PyObject *args) {
 
 static bool
 python_clear_predicate(void *data, void *key, unsigned keysz) {
-    PyObject *ret = PyObject_CallFunction(data, "y#", key, keysz);
+    PyObject *ret = PyObject_CallFunction(data, "y#", key, (Py_ssize_t)keysz);
     if (ret == NULL) {
         PyErr_Print();
         return false;
